pidor

+++ SchemAnalyser/assets/minecraft/shaders/fragment.glsl
@@ -1,10 +1,12 @@
 #version 330 core
+#include "minecraft:core.glsl"
+
 out vec4 FragColor;
 
 uniform vec3 color;
 uniform mat4 view;
-uniform double time_line;
+uniform float time_line;
 
-void main() { 
+void main() {
     FragColor = vec4(color,1.0); 
 }
